+++ lib/karafka/web/contracts/config.rb
@@ -97,8 +97,12 @@ module Karafka
           required(:per_page) { |val| val.is_a?(Integer) && val >= 1 && val <= 100 }
           required(:max_visible_payload_size) { |val| val.is_a?(Integer) && val >= 1 }
           required(:kafka) { |val| val.is_a?(Hash) }
-          required(:custom_css) { |val| val == false || (val.is_a?(String) && !val.empty?) }
-          required(:custom_js) { |val| val == false || (val.is_a?(String) && !val.empty?) }
+
+          nested(:custom) do
+            required(:css) { |val| val == false || (val.is_a?(String) && !val.empty?) }
+            required(:js) { |val| val == false || (val.is_a?(String) && !val.empty?) }
+            required(:nav_erb) { |val| val == false || (val.is_a?(String) && !val.empty?) }
+          end
 
           required(:dlq_patterns) do |val|
             val.is_a?(Array) &&